Update customer table and navigation links
Updated `Index.cshtml` to replace the `Password` column with `Address` and add a `PhoneNumber` column. Adjusted the table rows to display the new fields. Changed the title to "List of customer" and added a "Create New" link.

Modified `_Layout.cshtml` to update the "Customers" navigation link to use `asp-controller="Customers"` and `asp-action="Index"`.